:truck: Rename details component and refactor imports
