+++ src/app/core/messages.service.ts
@@ -1,10 +1,6 @@
 import { Injectable } from '@angular/core';
-import { Http, Headers } from '@angular/http';
-import { USER_ROUTES } from '@env/route';
-import { AuthService } from 'app/core/auth.service';
-import { UserEdit } from '../cube/user-profile/user-edit-information/user-edit-information.component';
-import { User } from '@cyber4all/clark-entity';
-import { MISC_ROUTES } from '../../environments/route';
+import { Http } from '@angular/http';
+import { MISC_ROUTES } from '@env/route';
 
 @Injectable()
 export class MessagesService {
