+++ src/app/cube/browse/browse.component.ts
@@ -1,9 +1,9 @@
 import { Observable, Subject, Subscription } from 'rxjs/Rx';
-import { SortType, OrderBy } from './../../shared/interfaces/query';
+import { SortType, OrderBy } from '../../shared/interfaces/query';
 import { ModalService, ModalListElement, Position} from '../../shared/modals';
 import { Router } from '@angular/router';
 import { LearningObject, AcademicLevel } from '@cyber4all/clark-entity/dist/learning-object';
-import { Component, OnInit, AfterViewChecked, OnDestroy, HostListener } from '@angular/core';
+import { Component, OnInit, OnDestroy, HostListener } from '@angular/core';
 import { LearningObjectService } from '../learning-object.service';
 import { ActivatedRoute } from '@angular/router';
 import { Query } from '../../shared/interfaces/query';
