+++ src/app/cube/details/children-detail-view/children-detail-view.component.scss
