+++ src/app/cube/details/details-content.component.scss
