+++ src/app/cube/details/details-content.component.ts
@@ -1,6 +1,4 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { EscapeHtmlPipe } from '../../../shared/pipes/keep-html.pipe';
-
 
 @Component({
     selector: 'learning-object-details-content',
