+++ src/app/cube/details/details.component.html
