+++ src/app/cube/details/details.component.scss
