+++ src/app/cube/details/details.component.spec.ts
