+++ src/app/cube/details/file-details/file-details.component.scss
