+++ src/app/cube/details/included-in/included-in.component.scss
