+++ src/app/cube/details/outcomes-detail-view/outcomes-detail-view.component.ts
