+++ src/app/cube/home/home.component.ts
@@ -1,6 +1,5 @@
-import { LearningObjectService } from './../learning-object.service';
+import { LearningObjectService } from '../learning-object.service';
 import { Component, OnInit } from '@angular/core';
-import { LearningObject } from '@cyber4all/clark-entity';
 import { Router } from '@angular/router';
 import { Query } from '../../shared/interfaces/query';
