+++ src/app/cube/user-profile/learning-objects.resolver.ts
@@ -1,15 +1,11 @@
-import { User } from '@cyber4all/clark-entity';
-import { AuthService } from 'app/core/auth.service';
+import { AuthService } from '../../core/auth.service';
 import {
   ActivatedRouteSnapshot,
   RouterStateSnapshot,
-  Resolve,
-  Router
+  Resolve
 } from '@angular/router';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs/Observable';
-import { USER_ROUTES } from '@env/route';
-import { UserService } from '../../core/user.service';
 import { LearningObjectService } from '../learning-object.service';
 
 @Injectable()
