+++ src/app/onion/content-upload/app/file-manager/upload-queue/upload-queue.component.ts
@@ -2,15 +2,14 @@ import {
   Component,
   OnInit,
   Input,
-  OnChanges,
   OnDestroy,
   Output,
   EventEmitter
 } from '@angular/core';
 import { BehaviorSubject, Subscription } from 'rxjs';
 import { File } from '@cyber4all/clark-entity/dist/learning-object';
-import { getPaths } from 'app/shared/filesystem/file-functions';
-import { getIcon } from 'app/shared/filesystem/file-icons';
+import { getPaths } from '../../../../../shared/filesystem/file-functions';
+import { getIcon } from '../../../../../shared/filesystem/file-icons';
 type LearningObjectFile = File;
 
 @Component({
