+++ src/app/onion/dashboard/dashboard.component.ts
@@ -1,13 +1,12 @@
 import { ModalService, ModalListElement, Position } from '../../shared/modals';
 import { NotificationService } from '../../shared/notifications';
-import { Component, OnInit, Input } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { Router, ActivatedRoute } from '@angular/router';
-import { Subject } from 'rxjs/Subject';
 import { LearningObjectService } from '../core/learning-object.service';
 import { LearningObject } from '@cyber4all/clark-entity';
 import { ChangeDetectorRef } from '@angular/core';
 import { TOOLTIP_TEXT } from '@env/tooltip-text';
-import { AuthService } from 'app/core/auth.service';
+import { AuthService } from '../../core/auth.service';
 
 
 @Component({
