+++ src/app/onion/dashboard/dashboard.resolver.ts
@@ -1,10 +1,6 @@
-import { User } from '@cyber4all/clark-entity';
-import { AuthService } from 'app/core/auth.service';
-import { ActivatedRouteSnapshot, RouterStateSnapshot, Resolve, Router } from '@angular/router';
+import { ActivatedRouteSnapshot, RouterStateSnapshot, Resolve } from '@angular/router';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs/Observable';
-import { USER_ROUTES } from '@env/route';
-import { UserService } from '../../core/user.service';
 import { LearningObjectService } from '../core/learning-object.service';
 
 @Injectable()
